in `kill` command, if the kill syscall fails due to an invalid signum - throw a regular SmashCmdError indicating that. If it fails for another reason - throw a SmashSysFailure, to print with perror

+++ Commands.cpp
@@ -341,6 +341,9 @@ pid_t KillCommand::execute() {
     }
 
     if ( kill(job->pid, this->sig_num) == -1 ) {
+        if (errno == EINVAL) {
+            throw SmashCmdError("kill: invalid arguments");
+        }
         throw SmashSysFailure("kill failed");
     }
